automation/requirements.txt
python-dotenv==1.0.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.125.0
requests==2.32.3
